refactor: Corrected file organization for reducers - Moved selectors for Auxillary sensors to the reducers index - Moved selectors for NetworkDispatch to the reducers index - Updated yarn.lock to accomodate new imports - Removed redux/ directory - Moved contents of redux/ to redux-features/
