+++ src/app/store.tsx
@@ -1,7 +1,4 @@
 import { configureStore, ThunkAction, Action, createImmutableStateInvariantMiddleware } from '@reduxjs/toolkit';
-import routerReducer from '../redux-features/Routing/RouterStore';
-import auxillarySensorsReducer from '../redux/AuxillarySensors/AuxillarySensors';
-import dispatchSystemsActionReducer from '../redux/NetworkDispatch/NetworkDispatch';
 
 //The one combined reducer containing all the smaller reducers to be used throughout the UI.
 import reducer from './reducersindex';
@@ -11,14 +8,8 @@ import reducer from './reducersindex';
  * Only allows for one reducer to be passed in to it.
  */
 export const store = configureStore({
-    reducer: {
-        router: routerReducer,
-        auxillarySensors: auxillarySensorsReducer,
-        dispatchSystemsAction: dispatchSystemsActionReducer,
-    },
-    middleware: [
-        createImmutableStateInvariantMiddleware({ ignore: ['router', 'auxillarySensors', 'networkDispatch'] }),
-    ],
+    reducer: reducer,
+    middleware: [createImmutableStateInvariantMiddleware({ ignore: ['router'] })],
 });
 
 export type RootState = ReturnType<typeof store.getState>;
