+++ src/redux-features/AuxillarySensors/AuxillarySensors.tsx
@@ -1,26 +1,25 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { RootState } from '../../app/store';
-import { ISocketMessageTransmittable } from '../../Models/Interfaces';
+//import { RootState } from '../../app/store';
 
-interface NetworkDispatchState {
-    dispatchSystemsAction: ((message: ISocketMessageTransmittable) => void) | undefined;
+interface AuxillarySensorsState {
+    currentTemperature: number;
 }
 
-const initialState: NetworkDispatchState = {
-    dispatchSystemsAction: undefined,
+const initialState: AuxillarySensorsState = {
+    currentTemperature: 0,
 };
 
 export const slice = createSlice({
-    name: 'networkDispatch',
+    name: 'auxillarySensors',
     initialState,
     reducers: {
-        setDispatchSystemsAction: (state, action: PayloadAction<(message: ISocketMessageTransmittable) => void>) => {
-            state.dispatchSystemsAction = action.payload;
+        setTemperature: (state, action: PayloadAction<number>) => {
+            state.currentTemperature = action.payload;
         },
     },
 });
 
-export const { setDispatchSystemsAction } = slice.actions;
+export const { setTemperature } = slice.actions;
 
 // The function below is called a thunk and allows us to perform async logic. It
 // can be dispatched like a regular action: `dispatch(incrementAsync(10))`. This
@@ -35,8 +34,6 @@ export const { setDispatchSystemsAction } = slice.actions;
 // The function below is called a selector and allows us to select a value from
 // the state. Selectors can also be defined inline where they're used instead of
 // in the slice file. For example: `useSelector((state: RootState) => state.counter.value)`
-export const selectDispatchSystemsAction = (
-    state: RootState,
-): ((message: ISocketMessageTransmittable) => void) | undefined => state.dispatchSystemsAction.dispatchSystemsAction;
+// export const selectTemperature = (state: RootState): number => state.auxillarySensors.currentTemperature;
 
 export default slice.reducer;
