+++ src/redux-features/NetworkDispatch/NetworkDispatch.tsx
@@ -1,25 +1,25 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { RootState } from '../../app/store';
+import { ISocketMessageTransmittable } from '../../Models/Interfaces';
 
-interface AuxillarySensorsState {
-    currentTemperature: number;
+interface NetworkDispatchState {
+    dispatchSystemsAction: ((message: ISocketMessageTransmittable) => void) | undefined;
 }
 
-const initialState: AuxillarySensorsState = {
-    currentTemperature: 0,
+const initialState: NetworkDispatchState = {
+    dispatchSystemsAction: undefined,
 };
 
 export const slice = createSlice({
-    name: 'auxillarySensors',
+    name: 'networkDispatch',
     initialState,
     reducers: {
-        setTemperature: (state, action: PayloadAction<number>) => {
-            state.currentTemperature = action.payload;
+        setDispatchSystemsAction: (state, action: PayloadAction<(message: ISocketMessageTransmittable) => void>) => {
+            state.dispatchSystemsAction = action.payload;
         },
     },
 });
 
-export const { setTemperature } = slice.actions;
+export const { setDispatchSystemsAction } = slice.actions;
 
 // The function below is called a thunk and allows us to perform async logic. It
 // can be dispatched like a regular action: `dispatch(incrementAsync(10))`. This
@@ -34,6 +34,8 @@ export const { setTemperature } = slice.actions;
 // The function below is called a selector and allows us to select a value from
 // the state. Selectors can also be defined inline where they're used instead of
 // in the slice file. For example: `useSelector((state: RootState) => state.counter.value)`
-export const selectTemperature = (state: RootState): number => state.auxillarySensors.currentTemperature;
+// export const selectDispatchSystemsAction = (
+// state: RootState,
+// ): ((message: ISocketMessageTransmittable) => void) | undefined => state.dispatchSystemsAction.dispatchSystemsAction;
 
 export default slice.reducer;
